refactor: split screenshot overlay helper
